pyproject.toml
[tool.poetry]
name = "paper-agents"
version = "0.1.0"
description = ""
authors = ["Tswoen <15071340+tswoen@user.noreply.gitee.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
requests = ">=2.32.5,<3.0.0"
python-dotenv = ">=1.1.1,<2.0.0"
pyyaml = ">=6.0.2,<7.0.0"
tenacity = ">=9.1.2,<10.0.0"
arxiv = ">=2.2.0,<3.0.0"
aiohttp = ">=3.12.15,<4.0.0"
langchain-community = ">=0.3.29,<0.4.0"
simhash = ">=2.1.2,<3.0.0"
pymupdf = ">=1.26.4,<2.0.0"
pyautogen = ">=0.2.20,<0.3.0"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "primary"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
